refactor(cf): relocate cf cells signal configs out of list framework
